Refactor: product-search entity 패키지를 product entity 패키지로 이동
