+++ build.gradle
@@ -118,6 +118,7 @@ project(':payment-api'){
 project(':product'){
     dependencies {
         implementation(project(':common'))
+        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     }
 }
