+++ product-search-api/src/main/java/org/ecommerce/productsearchapi/repository/SellerRepository.java
@@ -1,6 +1,6 @@
 package org.ecommerce.productsearchapi.repository;
 
-import org.ecommerce.productsearchapi.entity.SellerRep;
+import org.ecommerce.product.entity.SellerRep;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
